pyproject.toml
[project]
name = "OBESC"
version = "0.1.0"
description = "Projeto de extensão da UFSJ que permite localizar, filtrar e detalhar as atividades de extensão em um mapa da cidade de São João del-Rei."
requires-python = ">=3.14"
dependencies = [
    "folium>=0.20.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.54.0",
    "streamlit-folium>=0.26.1",
]
